+++ sql/init.sql
@@ -1,6 +1,6 @@
 -- TABLE "users" --
 CREATE TABLE IF NOT EXISTS users (
-  "id"        BIGINT UNIQUE NOT NULL PRIMARY KEY,
+  "id"        INT UNIQUE NOT NULL PRIMARY KEY,
   "firstname" CITEXT NOT NULL,
   "lastname"  CITEXT NOT NULL,
   "age"       INT,
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS users (
 -- TABLE "orders" --
 CREATE TABLE IF NOT EXISTS orders (
   "id"            SERIAL UNIQUE PRIMARY KEY,
-  "user_id"       BIGINT NOT NULL,
+  "user_id"       INT NOT NULL,
   "firstname"     CITEXT DEFAULT NULL,
   "lastname"      CITEXT DEFAULT NULL,
   "phone"         CITEXT DEFAULT NULL,
@@ -22,6 +22,6 @@ CREATE TABLE IF NOT EXISTS orders (
 
 -- TABLE "states" --
 CREATE TABLE IF NOT EXISTS states (
-  "user_id"     BIGINT UNIQUE PRIMARY KEY NOT NULL,
+  "user_id"     INT UNIQUE PRIMARY KEY NOT NULL,
   "state"       INT DEFAULT 0 NOT NULL
 );
